+++ src/components/common/label/Label.stories.tsx
